Removed view model
